Code refractoring
