+++ pythonfmu/__init__.py
@@ -1,1 +1,4 @@
-from pythonfmu._version import __version__
+from ._version import __version__
+from .enums import Fmi2Causality, Fmi2Initial, Fmi2Variability
+from .variables import Boolean, Integer, Real, String
+from .fmi2slave import Fmi2Slave
